requirements.txt
# Core Dependencies
anthropic>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.7
typer>=0.9.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1
pytest-xdist>=3.3.1
pytest-timeout>=2.2.0
pytest-env>=1.1.3
pytest-sugar>=0.9.7
pytest-randomly>=3.15.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0
bandit>=1.7.5
safety>=2.3.5
pre-commit>=3.5.0
pylint>=3.0.2
radon>=6.0.1

# Type Hints
types-requests>=2.31.0.2
types-python-dateutil>=2.8.19.14
types-setuptools>=69.0.0.0
types-PyYAML>=6.0.12.12

# API and Web
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=9.0,<11.0
aiohttp>=3.9.1
gunicorn>=21.2.0
starlette>=0.27.0
pydantic>=2.0.0
pydantic-settings==2.0.3
email-validator>=2.1.0.post1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
openpyxl>=3.1.2
xlrd>=2.0.1
pyarrow>=14.0.1
polars>=0.20.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=41.0.7
pyjwt>=2.8.0
oauthlib>=3.2.2
requests-oauthlib>=1.3.1

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
redis>=4.5.0
aioredis>=2.0.1
motor>=3.3.2
pymongo>=4.6.1

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.1.0
sentry-sdk>=1.39.1
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-exporter-otlp>=1.22.0

# Development Tools
ipython>=8.18.0
jupyter>=1.0.0
jupyterlab>=4.0.9
notebook>=7.0.6
debugpy>=1.8.0
ipdb>=0.13.13
watchdog>=3.0.0

# GIS Dependencies
geopandas>=0.14.1
shapely>=2.0.2
pyproj>=3.6.1
rasterio>=1.3.9
fiona>=1.9.5
folium>=0.15.1
geopy>=2.4.1

# Financial Dependencies
yfinance>=0.2.33
pandas-ta>=0.3.14b0
ccxt>=4.1.87
python-binance>=1.0.19
alpaca-trade-api>=3.0.2

# Social Network Dependencies
networkx>=3.2.1
scipy>=1.12.0
scikit-learn>=1.2.0
nltk>=3.8.1
spacy>=3.7.2
transformers>=4.30.0
torch>=2.0.0

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.3
mkdocs-material>=9.4.14

# HTTP Client
httpx==0.25.1

# Added from the code block
GDAL==3.6.2
statsmodels>=0.14.0
prophet>=1.1.4
pmdarima>=2.0.3
openai>=1.3.0
cohere>=4.37
google-generativeai>=0.3.0
replicate>=0.22.0
huggingface-hub>=0.19.0

# New dependencies
twilio
python-gnupg
tweepy
yagmail
